Split up headers and source, move headers into namespace
