+++ headers/mdi/MDICommon.h
